refactor(security): move the destination host check out of batch_exports
The SSRF check that guards a customer-supplied batch export or warehouse-import
target lived in batch_exports, so core's Postgres integration and
managed_migrations reached across a product boundary to import it. That coupling
blocks isolating batch_exports behind a facade.

Move resolve_and_validate_host, resolve_and_validate_url and their helpers into
posthog/security/url_validation.py unchanged, and repoint all three callers. The
only edit is urlparse -> urlparse.urlparse, because the destination module imports
urllib.parse under that alias.

Behavior is identical. The check still duplicates the concern is_url_allowed
already covers in the same module; unifying them changes what we reject, so that
lands on top of this.

Generated-By: PostHog Desktop
Task-Id: 0d94ba3e-d686-444e-bd78-a0e9cac6092c
